n64: Move RODATA out of RAM
The Nintendo64 has a very slow but large ROM (the cartridge).  Typically
the whole binary is loaded into RAM for faster instruction fetching and
making use if instruction cache.  In Go this includes large data
structures like pclntab and also files embedded via the embed package.
To free up RAM and allow binaries larger than the RAM, this patch does
two things:
- Move the RODATA segment after the DATA segment, both in virtual
  address space and in the binary.  This is not necessary but makes
  the initial loading of TEXT and DATA into RAM a single DMA copy.
- Offset the virtual address of the RODATA segment from RAM to
  cartridge.

Be aware that this makes reading from RODATA slower because (1) it's
read over the PI bus and (2) caching is not possible.