Fix interrupt exception with no pending interrupts
The VR4300 User Manual (p. 199) says "The IP field of the Cause register
indicates current interrupt requests. It is possible before this
register is read [...] that more than one of the bits can be
simultaneously cleared if the interrupt request signal is deasserted."

This is what happened with the SummerCart64 v2.18.1 button interrupt,
which seems to have a delay until the interrupt is cleared.